+++ src/adk_docker_uv/utils/__init__.py
@@ -1,5 +1,6 @@
 """Utility modules."""
 
+from adk_docker_uv.utils.env_parser import parse_json_list_env
 from adk_docker_uv.utils.log_config import setup_file_logging
 
-__all__ = ["setup_file_logging"]
+__all__ = ["parse_json_list_env", "setup_file_logging"]